+++ controllers/utility.go
@@ -10,7 +10,7 @@ import (
 
 const (
 	ignoreClusterDeletion = "alpha.giantswarm.io/ignore-cluster-deletion"
-	keepValid             = "keep-valid"
+	keepUntil             = "keep-until"
 
 	// defaultTTL is the default time to live for a cluster.
 	defaultTTL = 8 * time.Hour
@@ -18,8 +18,8 @@ const (
 	// eventDefaultTTL is the default time when we sent a `ClusterMarkedForDeletion` event.
 	eventDefaultTTL = defaultTTL - 1*time.Hour
 
-	// keepValidTimeLayout is the layout for the `keep-valid` label.
-	keepValidTimeLayout = "2006-01-02"
+	// keepUntilTimeLayout is the layout for the `keep-until` label.
+	keepUntilTimeLayout = "2006-01-02"
 )
 
 func requeue() reconcile.Result {