+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Consider cluster deletion when `keep-valid` label is set.
+- Consider cluster deletion when `keep-until` label is set.
 
 ## [0.1.0] - 2021-12-07